Remove unused Groovy script plugins
Strangely, it's not possible to remove those related License Report. When I did remove them, it failed the `PluginTest` because of not being able to locate `com.github.jk1.dependency-license-report` plugin of version `2.0`.

The plugin is added in `buildScr/build.gradle.kts` (see the `plugins { ... }` section). The `PluginTest` uses `buildScr` copied to the test project.