Improve the method name
Sometimes we need to call this method under the `run` clausure of the action. This helps in debugging because we can set a breakpoint on the call. At the same time, it allows avoiding unnecessary variables.

So the name `render` is better than `run` even if we speak `Command` pattern in this case.